Move to data_amodel namespace
