+++ wikidatarefisland/__init__.py
@@ -1,4 +1,3 @@
 from .config import Config
-from .schemaorg_normalizer import SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode
 
-__all__ = [Config, SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode]
+__all__ = [Config]
