+++ wikidatarefisland/data_model/__init__.py
@@ -1,0 +1,3 @@
+from .schemaorg_normalizer import SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode
+
+__all__ = [SchemaOrgNormalizer, SchemaOrgGraph, SchemaOrgNode]
